:bug: fix: Restringir deleção e modificação apenas aos criadores

+++ src/answers/answers.service.ts
@@ -20,13 +20,14 @@ export class AnswersService {
         data: {
           response: createResponseDto.response,
           question_id: createResponseDto.question_id,
+          user_id: createResponseDto.user_id,
         },
       });
       return newResponse;
     } catch (error) {
       if (error.code === 'P2003') {
         throw new HttpException(
-          'Pergunta não encontrada.',
+          'Pergunta ou usuário não encontrados.',
           HttpStatus.BAD_REQUEST,
         );
       } else {
@@ -41,17 +42,27 @@ export class AnswersService {
   async updateResponse(
     id: number,
     updateResponseDto: UpdateResponseDto,
+    user_id_request: number,
   ): Promise<Response> {
     try {
-      const existingResponse = await this.prisma.questions.findUnique({
+      const existingResponse = await this.prisma.answers.findUnique({
         where: { id },
       });
+
       if (!existingResponse) {
         throw new HttpException(
           'Resposta não encontrada.',
           HttpStatus.NOT_FOUND,
         );
       }
+
+      if (existingResponse.user_id !== user_id_request) {
+        throw new HttpException(
+          'Resposta pertence a outro usuário.',
+          HttpStatus.UNAUTHORIZED,
+        );
+      }
+
       const updateResponse = await this.prisma.answers.update({
         where: { id },
         data: {
@@ -61,12 +72,17 @@ export class AnswersService {
 
       return updateResponse;
     } catch (error) {
+      if (error instanceof HttpException) {
+        throw error;
+      }
+
       if (error.code === 'P2003') {
         throw new HttpException(
           'Pergunta não encontrada',
           HttpStatus.BAD_REQUEST,
         );
       }
+
       throw new HttpException(
         'Erro ao atualizar a resposta.',
         HttpStatus.INTERNAL_SERVER_ERROR,
@@ -74,12 +90,23 @@ export class AnswersService {
     }
   }
 
-  async deleteResponse(id: number): Promise<{ message: string }> {
+  async deleteResponse(
+    id: number,
+    user_id_request: number,
+  ): Promise<{ message: string }> {
     const existingResponse = await this.prisma.answers.findUnique({
       where: { id },
     });
+
     if (!existingResponse) {
       throw new HttpException('Resposta não encontrada.', HttpStatus.NOT_FOUND);
+    }
+
+    if (existingResponse.user_id !== user_id_request) {
+      throw new HttpException(
+        'Resposta pertence a outro usuário.',
+        HttpStatus.UNAUTHORIZED,
+      );
     }
 
     await this.prisma.answers.delete({
